+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](semver.md).
 ### Fixed
 ### Security
 
+## [1.1.0] (2020-09-01)
+
+### Added
+* Support for tar, gz, bz2 and xz files
+* Support to add a complete directory with -d
+
+### Fixed
+* Exclude the output file from all inputs
+
 ## [1.0.0] (2020-09-01)
 
 ### Added